requirements.txt
streamlit
httpx
python-dotenv
openai
asyncio